NTR - Fix storefront language switch
The storefront snippet translation system works with
different domains for different languages.

That means that we can just display the languages
that have an domain configured for this salesChannel.

Previously we could choose all languages that are configured
for that salesChannel, but if the domain was missing for that
language we got an LanguageNotFoundException.

Reproducable by changing the storefront to german after a
`psh install`.