NEXT-5428 - Fix CSRF for ajax loaded templates
The content-type of all ajax requests was set to "application/json" as default.
The content-type of the response was never overwritten, because of that the
content-type of the request was used, hence the CSRF tokens weren't replaced
as it only works on content-type "text/html".
The solution was to set the content-type explicitly to "text/html" every time
we render a storefront template.